Update some tests.
