+++ packages/types/src/workspace-config.ts
@@ -13,8 +13,6 @@ export interface NotifierConfig {
 export interface RunnerConfig {
 	archivableTargets: string[];
 	cacheLifetime: string;
-	implicitDeps: string[];
-	implicitInputs: string[];
 	inheritColorsForPipedTasks: boolean;
 	logRunningCommand: boolean;
 }
